Update Manage Category to let the possibilty for the user to switch during the process to add in DB
