+++ my_qcm/app/src/main/java/com/tactfactory/my_qcm/data/webservice/CategoryWSAdapter.java
@@ -1,7 +1,10 @@
 package com.tactfactory.my_qcm.data.webservice;
 
 
+import android.content.Context;
+import android.os.AsyncTask;
 import android.support.annotation.NonNull;
+import android.widget.ArrayAdapter;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
@@ -11,7 +14,9 @@ import com.loopj.android.http.AsyncHttpResponseHandler;
 import com.loopj.android.http.JsonHttpResponseHandler;
 import com.loopj.android.http.RequestParams;
 import com.loopj.android.http.TextHttpResponseHandler;
+import com.tactfactory.my_qcm.R;
 import com.tactfactory.my_qcm.configuration.MyQCMConstants;
+import com.tactfactory.my_qcm.data.CategSQLiteAdapter;
 import com.tactfactory.my_qcm.entity.Categ;
 
 import org.json.JSONArray;
@@ -28,6 +33,7 @@ import java.util.Date;
 import java.util.Iterator;
 import java.util.List;
 import java.util.ListIterator;
+import java.util.concurrent.ExecutionException;
 
 import cz.msebera.android.httpclient.Header;
 
@@ -36,8 +42,13 @@ public class CategoryWSAdapter {
 
     String response;
     MyQCMConstants myQCMConstants;
+    Context context;
 
-    public void getCategoryRequest (Integer user_id, String url, final CallBack callback) {
+    public CategoryWSAdapter(Context context) {
+        this.context = context;
+    }
+
+    public void getCategoryRequest (Integer user_id, String url) {
         AsyncHttpClient asyncHttpClient = new AsyncHttpClient();
         asyncHttpClient.setConnectTimeout(60000);
         asyncHttpClient.setTimeout(600000);
@@ -55,8 +66,7 @@ public class CategoryWSAdapter {
                 for(Categ categ:categories) {
                     System.out.println("On success = " + categ.getName());
                 }
-
-                callback.methods(categories);
+                ManageCategDB(categories);
             }
 
             @Override
@@ -64,7 +74,6 @@ public class CategoryWSAdapter {
                 response = responseString;
                 ArrayList<Categ> categories = new ArrayList<Categ>();
                 System.out.println("On failure");
-                callback.methods(categories);
             }
 
             @Override
@@ -78,14 +87,11 @@ public class CategoryWSAdapter {
                 }
                 System.out.println("On failure = " + str);
                 response = "false";
-                callback.methods(categories);
+
             }
         });
     }
 
-    public interface CallBack{
-        void methods(ArrayList<Categ> categories);
-    }
 
     public  ArrayList<Categ> responseToList(String response)
     {
@@ -104,6 +110,86 @@ public class CategoryWSAdapter {
         return categories;
 
 
+    }
+    public void ManageCategDB (ArrayList<Categ> response)
+    {
+        if (response.isEmpty() == false) {
+            // get the list of categ in Flux and add on listView
+            ArrayList<Categ> list = response;
+            ArrayList<String> listResult = null;
+
+            // Call the AsyncTask to add Categ on the DB and returns the list of result
+            try {
+                listResult = new ManageCategDBTask().execute(list).get();
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            } catch (ExecutionException e) {
+                e.printStackTrace();
+            }
+
+        }
+    }
+
+    /**
+     * Asyntask To Manage Categories get on the Flow in the DB
+     */
+    public class ManageCategDBTask extends AsyncTask<ArrayList<Categ>,Void,ArrayList<String>> {
+
+        @Override
+        protected ArrayList<String> doInBackground(ArrayList<Categ>... params) {
+            ArrayList<Categ> categories =  new ArrayList<Categ>();
+            ArrayList<String> results = new ArrayList<String>();
+            categories = params[0];
+
+            CategSQLiteAdapter categSQLiteAdapter = new CategSQLiteAdapter(context);
+            categSQLiteAdapter.open();
+            ArrayList<Categ> categoriesDB = categSQLiteAdapter.getAllCateg();
+
+            for(Categ categ : categories)
+            {
+                Categ tempCateg ;
+                //Try to find a Categ with this id_server
+                tempCateg = categSQLiteAdapter.getCategById_server(categ.getId_server());
+
+                //If Categ not exist on Mobile DB
+                if(tempCateg == null)
+                {
+                    //Add categ on the DB
+                    long result = categSQLiteAdapter.insert(categ);
+                    results.add(String.valueOf(result));
+                }
+                else
+                {
+                    long result = categSQLiteAdapter.update(categ);
+                    results.add(String.valueOf(result));
+                }
+
+            }
+            //delete check is existe on the DB but not
+            if(categoriesDB != null) {
+                for (Categ categorie : categoriesDB) {
+                    Boolean isExist = false;
+                    for (Categ categ : categories) {
+                        if (categ.getId_server() == categorie.getId_server()) {
+                            isExist = true;
+                        }
+                    }
+
+                    if (isExist == false) {
+                        long result = categSQLiteAdapter.delete(categorie);
+                    }
+                }
+            }
+            categSQLiteAdapter.close();
+
+
+            return results;
+        }
+
+        @Override
+        protected void onPostExecute(ArrayList<String> result) {
+            super.onPostExecute(result);
+        }
     }
